backend/requirements.txt
Django==4.2
djangorestframework
psycopg2-binary
openai
requests
google-genai
django-cors-headers




